dsl/fgl/requirements.txt
antlr4-python3-runtime==4.11.1
antlr4-tools==0.1
graphviz==0.20.1
tabulate==0.9.0
coverage==7.1.0
uvicorn==0.20.0
contextvars==2.4


